Create esp32_web_server_STA_mode.ino
In the Land of Code:

Once upon a time, in the land of code, there lived a magical ESP32 device. This device had the power to connect to the internet and listen for messages from other devices. Let's embark on a magical journey through this code:

🏰 Our adventure began in the setup area. We decided to connect to a Wi-Fi network, but we didn't want to reveal our secret name and password. While we waited for the Wi-Fi network to acknowledge our presence, we chanted, "Trying to Connect!" every 2 seconds, so everyone would know we were making new friends. When the Wi-Fi network finally accepted our friendship, we exclaimed, "Connected! Good to Go!" and proudly announced our magical address.

🧙‍♂️ With our new Wi-Fi friend by our side, we decided to create a magical server on port 80. This server allowed us to receive messages from other devices on the internet.

In our daily adventures (in the loop), we kept an eye on our magical mailbox (server.available()). Whenever a new friend knocked on our door (connected to our server), we greeted them warmly and invited them inside (we set up a client to communicate with them).

As our new friend started sending us messages (characters), we listened carefully and collected each character in a special box (request) to form a complete message.

We watched for a magical "newline" character, which signaled the end of a message. When we saw it, we couldn't contain our excitement! We shared the entire message with the world (printed it to the serial monitor). We even hinted, "You can add your special magic here to do things with the message."

After sharing the message, we cleared our special box (request) to make room for the next message.

When our friend finished sharing their messages with us, we said our farewells and closed the door (stopped the client connection), knowing that we had completed another day of magical adventures.

And that, my dear friend, is the enchanting tale of how this magical ESP32 device listened to messages from other devices on the internet and shared those messages with the world! 🌟🏰🚀